chore: 1b.3.c clippy cleanups
- io_other_error: zellij_input_writer.rs uses io::Error::other
- unnecessary_cast: zellij_pane.rs drops size.rows/cols `as usize`
- B1 review: drop no-op Server::new() in writer test + tighten
  comment block

+++ mux/src/zellij_pane.rs
@@ -186,8 +186,8 @@ impl Pane for ZellijPane {
         // Notify zellij-server so it recomputes pane geometry.
         let msg = zellij_utils::ipc::ClientToServerMsg::TerminalResize {
             new_size: zellij_utils::pane_size::Size {
-                rows: size.rows as usize,
-                cols: size.cols as usize,
+                rows: size.rows,
+                cols: size.cols,
             },
         };
         self.input_sender
